use size_t type for memory amount in memory allocation routines

+++ bld/as/c/asitest.c
@@ -96,8 +96,8 @@ void AsmWarning( char *msg ) {
     printf( "AsmWarning (line %d): %s\n", curLine, msg );
 }
 
-void *AsmAlloc( unsigned amount ) {
-//*********************************
+void *AsmAlloc( size_t amount ) {
+//*******************************
 // CC provides this
 
     return( TRMemAlloc( amount ) );
